requirements.txt
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
python-telegram-bot==22.3
ultralytics==8.3.161
